idx: move Index into separate file
